+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Utility class used to load plugins from shared libraries (#1035, **@RiscadoA**).
 - Implemented metrics and profiling tools (#1150, **@roby2014**).
 - GL timer for profiling sections of rendering code (#1228, **@tomas7770**).
+- Shadows plugin (#1185, **@tomas7770**).
 
 ### Changed